Moved Text and Font out into their own headers
